added service package and some refactoring
